march can now dump its triangles so you can view them.
if DUMP_TRIANGLES is #defined in march.c, march will dump a binary
file containing all the vertex data for the triangles it found.
This commit adds bin_to_obj, a utility that will convert this binary
file to .obj format for use in 3D mesh viewers.  This allows one
to sanity-check the march code or visualize the scalar field being
polygonized.